requirements.txt
sqlparse
djangorestframework
django
Pillow
mysqlclient
psycopg
psycopg2
django-filter